refactor(web): extract file-change-tracker from SessionDetail
Move FileChangeTracker (+ private FileTrackerSection/Item) and
getFileTrackerItemCount out of the 1840-line SessionDetail.tsx into
session-detail/file-change-tracker.tsx. Pure move + import rewire, no
behavior change. First sub-domain of CS-13.
